backend/src/shared/requirements.txt
langchain-community==0.2.16
python-dotenv==1.0.1
boto3==1.35.13
langchain-openai==0.1.23
langchain-anthropic==0.1.23


